Removed straem reduce interfaces that colapses the whole stream into a single value
