+++ tests/CMakeLists.txt
@@ -4,25 +4,25 @@ include_directories( ${PROJECT_SOURCE_DIR} )
 # Set tests
 set( TESTS_CPU_SEQ
      pipeline1              pipeline2              farm1
-     farm2                  farm3                  stream_reduce1
+     farm2                  farm3                  
      pipeline+farm          map1                   #map2
      map3                   reduce1                reduce2
      stencil1               mapreduce1             divideandconquer1
-     divideandconquer-fibonacci stream_reduce_new
+     divideandconquer-fibonacci stream_reduce_example
      stream_iteration1
      stream_iteration_pipe
 #     PPATtest               PPATfarm
 # Examples in the Deliverable
      pipeline_example       farm_example           filter_example  
-     stream_reduce_example  pipeline+filter_test_ppi_example
+     pipeline+filter_test_ppi_example
 )
 
 set( TESTS_CPU_OMP
      pipeline1              pipeline2              farm1
-     farm2                  farm3                  stream_reduce1
+     farm2                  farm3                  
      pipeline+farm          map1                   #map2
      map3                   reduce1                reduce2
-     stencil1               divideandconquer1      stream_reduce_new
+     stencil1               divideandconquer1      
  
      divideandconquer-fibonacci
 # Examples in the Deliverable
@@ -32,8 +32,8 @@ set( TESTS_CPU_OMP
 
 set( TESTS_CPU_TBB                           
      pipeline1              pipeline2              farm1
-     farm2                  farm3                  stream_reduce1 
-     pipeline+farm          map1                   stream_reduce_new
+     farm2                  farm3                   
+     pipeline+farm          map1                   
      reduce1                reduce2
      stencil1 divideandconquer1
      divideandconquer-fibonacci
@@ -44,11 +44,11 @@ set( TESTS_CPU_TBB
 
 set( TESTS_CPU_THR                           
      pipeline1              pipeline2              farm1
-     farm2                  farm3                  stream_reduce1
+     farm2                  farm3                  
      pipeline+farm          map1                   #map2
      map3                   reduce1                reduce2
      stencil1               divideandconquer1      mapreduce1
-     divideandconquer-fibonacci stream_reduce_new
+     divideandconquer-fibonacci
      stream_iteration1
      stream_iteration_pipe
 # Examples in the Deliverable
